+++ devicemanagement/CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `getModulePCIENumPortsMaxSpeed` test renamed to `getModulePCIEPortsMaxSpeed`.
 - Disabled getModulePCIEPortsMaxSpeed test on SysEMU.
 - Disabled setPCIELinkSpeed test on Silicon. To be enabled back in SW-13272
+- Disable active power management before exiting test setModuleActivePowerManagement.
 ### Deprecated
 ### Removed
 - Workaround for SP Trace